count using repository method

+++ be/src/main/java/com/brodeckyondrej/SignUp/business/dto/category/CategoryGetListDto.java
@@ -7,10 +7,10 @@ import java.util.UUID;
 
 @Getter
 public class CategoryGetListDto extends NamedDtoWithId {
-    private final int numberOfSigns;
+    private final long numberOfSigns;
     private final NamedDtoWithId subjectNameId;
 
-    public CategoryGetListDto(UUID id, String name, int numberOfSigns, NamedDtoWithId subjectNameId) {
+    public CategoryGetListDto(UUID id, String name, long numberOfSigns, NamedDtoWithId subjectNameId) {
         super(id, name);
         this.numberOfSigns = numberOfSigns;
         this.subjectNameId = subjectNameId;
